+++ src/pages/manageProjects/newProject/NewProjectDialog.jsx
@@ -1,39 +1,11 @@
 import { useEffect, useState, useMemo } from 'react'
 import { Dialog } from 'primereact/dialog'
-import { Dropdown } from 'primereact/dropdown'
 import { toast } from 'react-toastify'
 
 import axios from 'axios'
 import { Button, Spacer, InputText, Toolbar } from '@ynput/ayon-react-components'
 import SettingsEditor from '/src/containers/settingsEditor'
-import { loadAnatomyPresets } from '/src/utils'
-
-const PresetDropdown = ({ selectedPreset, setSelectedPreset }) => {
-  const [presetList, setPresetList] = useState([])
-
-  useEffect(() => {
-    loadAnatomyPresets().then((r) => {
-      setPresetList(r)
-      if (!selectedPreset) {
-        setSelectedPreset(r[0].name)
-      }
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  return (
-    <Dropdown
-      value={selectedPreset}
-      onChange={(e) => setSelectedPreset(e.value)}
-      options={presetList}
-      optionValue="name"
-      optionLabel="title"
-      tooltip="Preset"
-      tooltipOptions={{ position: 'bottom' }}
-      style={{ minWidth: 200 }}
-    />
-  )
-}
+import PresetDropdown from './PresentDropdown'
 
 const NewProjectDialog = ({ onHide }) => {
   const [schema, setSchema] = useState(null)
